Open the Add Item chooser from the empty state in type-agnostic views.
