+++ src/action/containers/AppHeaderContainer/index.tsx
@@ -1,56 +1,17 @@
-import type { ComponentType, SVGProps } from 'react'
+import { AUTHENTICATOR_ENABLED } from '@tetherto/pearpass-lib-constants'
 
-import { t } from '@lingui/core/macro'
-import {
-  AUTHENTICATOR_ENABLED,
-  UNSUPPORTED
-} from '@tetherto/pearpass-lib-constants'
-import {
-  ContextMenu,
-  NavbarListItem,
-  useTheme
-} from '@tetherto/pearpass-lib-ui-kit'
-import {
-  Key,
-  TwoFactorAuthenticationOutlined
-} from '@tetherto/pearpass-lib-ui-kit/icons'
-
+import { AddItemContextMenu } from '../AddItemContextMenu'
 import { AppHeaderAddItemTrigger, AppHeaderV2 } from '../AppHeaderV2'
 import { ImportItemOrVaultModalContentV2 } from '../../../shared/containers/ImportItemOrVaultModalContentV2'
 import { useAppHeaderContext } from '../../../shared/context/AppHeaderContext'
 import { useModal } from '../../../shared/context/ModalContext'
 import { useRouter } from '../../../shared/context/RouterContext'
-import { useRecordMenuItemsV2 } from '../../../shared/hooks/useRecordMenuItemsV2'
 import { isFavorite } from '../../../shared/utils/isFavorite'
-import { useCreateOrEditRecord } from '../../hooks/useCreateOrEditRecord'
-
-const ADD_MENU_WIDTH = 220
-
-const PASSWORD_TYPE = '__password__'
-const AUTHENTICATOR_CODE_TYPE = '__authenticator__'
-
-type AddMenuExtraItem = {
-  type: string
-  label: string
-  Icon: ComponentType<SVGProps<SVGSVGElement>>
-}
 
 export const AppHeaderContainer = () => {
-  const { theme } = useTheme()
-  const {
-    currentPage,
-    state: routerState,
-    navigate
-  } = useRouter() as {
+  const { currentPage, state: routerState } = useRouter() as {
     currentPage: string
     state: { recordType?: string; folder?: string } | undefined
-    navigate: (
-      page: string,
-      opts?: {
-        params?: Record<string, unknown>
-        state?: Record<string, unknown>
-      }
-    ) => void
   }
   const {
     searchValue,
@@ -60,9 +21,7 @@ export const AppHeaderContainer = () => {
     isSidebarCollapsed,
     setIsSidebarCollapsed
   } = useAppHeaderContext()
-  const { defaultItems } = useRecordMenuItemsV2()
   const { setModal } = useModal()
-  const { handleCreateOrEditRecord } = useCreateOrEditRecord()
 
   if (currentPage !== 'vault') {
     return null
@@ -76,38 +35,6 @@ export const AppHeaderContainer = () => {
   const selectedFolder =
     routerState?.folder && !isFavoritesView ? routerState.folder : undefined
 
-  const passwordItem: AddMenuExtraItem = {
-    type: PASSWORD_TYPE,
-    label: t`Password`,
-    Icon: Key
-  }
-
-  const authenticatorItem: AddMenuExtraItem = {
-    type: AUTHENTICATOR_CODE_TYPE,
-    label: t`Authenticator Code`,
-    Icon: TwoFactorAuthenticationOutlined
-  }
-
-  const handleSelectType = (type: string) => {
-    setIsAddMenuOpen(false)
-
-    if (type === PASSWORD_TYPE) {
-      handleCreateOrEditRecord({ recordType: 'password' })
-      return
-    }
-
-    if (type === AUTHENTICATOR_CODE_TYPE) {
-      navigate('authenticator', { params: {} })
-      return
-    }
-
-    handleCreateOrEditRecord({
-      recordType: type,
-      selectedFolder,
-      isFavorite: isFavoritesView || undefined
-    })
-  }
-
   const handleImportClick = () => {
     setModal(<ImportItemOrVaultModalContentV2 />)
   }
@@ -116,26 +43,12 @@ export const AppHeaderContainer = () => {
     setIsSidebarCollapsed((value) => !value)
   }
 
-  const renderMenuItem = (item: {
-    type: string
-    label: string
-    Icon: ComponentType<SVGProps<SVGSVGElement>>
-  }) => (
-    <NavbarListItem
-      key={item.type}
-      size="small"
-      label={item.label}
-      testID={`add-menu-${item.type}`}
-      icon={<item.Icon color={theme.colors.colorTextPrimary} />}
-      onClick={() => handleSelectType(item.type)}
-    />
-  )
-
   const addItemControl = (
-    <ContextMenu
-      open={isAddMenuOpen}
+    <AddItemContextMenu
+      isOpen={isAddMenuOpen}
       onOpenChange={setIsAddMenuOpen}
-      menuWidth={ADD_MENU_WIDTH}
+      selectedFolder={selectedFolder}
+      isFavoritesView={isFavoritesView}
       testID="app-header-add-menu"
       trigger={
         <AppHeaderAddItemTrigger
@@ -143,30 +56,7 @@ export const AppHeaderContainer = () => {
           onClick={() => setIsAddMenuOpen(!isAddMenuOpen)}
         />
       }
-    >
-      {defaultItems.map((item) =>
-        renderMenuItem({
-          type: item.type,
-          label: item.label,
-          Icon: item.OutlinedIcon
-        })
-      )}
-      {renderMenuItem(passwordItem)}
-      {UNSUPPORTED && (
-        <>
-          <div
-            style={{
-              height: 1,
-              backgroundColor: theme.colors.colorBorderPrimary,
-              margin: 0
-            }}
-            role="separator"
-            aria-hidden="true"
-          />
-          {renderMenuItem(authenticatorItem)}
-        </>
-      )}
-    </ContextMenu>
+    />
   )
 
   return (
